JVMCBC-5: Implement HELLO Command.
Motivation
----------
The HELLO command can be used to negotiate certain features between server and client.
For now JSON and compression are supported for datatypes. Also, a config setting is
exposed which allows to manually disable and change thresholds.

Modifications
-------------
A new handler has been added to the pipeline which handles the HELLO negotation. If supported
it passes an event through the pipeline that gets picked up by the codec to influence its
behavior.

Result
------
HELLO with JSON and compression are now supported and configurable.

Change-Id: I0f61df3de10b9c9634e4130c0e0ed98520453073