JVMCBC-32: Also convert partitionHosts to InetAddrs.
Motivation
----------
If the partition hosts are kept as strings and they are compared as hostnames,
issues can come up when ip addrs and hostnames are mixed and matched in the
configurations.

Modifications
-------------
By also converting them to InetAddrs and comparing them, the actual check is
delegated to the InetAddr implementation, which accounts for those issues
with proper checks.

Result
------
The core is now more resilient to ip/hostname mixes and able to resolve,
especially in the binary handler.

Change-Id: I3a91ee24049d70aafbee0501ab7ba7a877384f11