JVMCBC-38: Always subscribe on the computation scheduler.
This change makes sure that the observable the core returns is
never subscribed on the io event loops or other internal parts.
Instead, the computation scheduler is utilized which should have
enough power for downstream computations.

Note that if blocking operations are performed, it is expected that
downstream the subscription is switched to the IO scheduler!

Change-Id: I2d5785f9b46f5b5587fd62521663d932c5b1b381
Reviewed-on: http://review.couchbase.org/41481
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>