JVMCBC-35: Expose View retry codes in response status
Motivation
----------
Exposing the actual status code including the reason phrase from the
HTTP response is vital in the handling code so that it can make informed
decisions about wheter to retry or not.

Modifications
-------------
Both the actual status code and the reason phrase are exposed in the response
in order to allow better decisions in the calling code.

Result
------
Better error handling capabilities at the caller side.

Change-Id: I90128e919fee103952729c1b04f75c029bf4ac95
Reviewed-on: http://review.couchbase.org/42112
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>