JVMCBC-44 : improve the check for closed buckets
Motivation
-----------
After the first JVMCBC-44 refactor (see initial commit of BucketClosedTest),
some NPE arise because the check is too strong. Additional types of request
must be allowed to pass through.

Modifications
-------------
Introduced a marker interface to mark requests that can pass despite bucket conf
being either empty or not containing the target bucket.
The RequestHandler's check now filters on that interface, BootstrapMessage.

If the request don't pass the check, its Observable is failed with a
BucketClosedException.

Result
------
Any request to do some work with a closed bucket immediately fails into a
BucketClosedException (passed in the onError method of the request's Observable).
No more NullPointerExceptions in the RequestHandler's onEvent method.

Change-Id: Ie53f4e0f7fee2265566da03a52baaeec0d01c22a
Reviewed-on: http://review.couchbase.org/42230
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>