JVMCBC-45: Give config loading a chance to be retried.
Motivation
----------
Lots of exceptions have been reported where the GetBucketConfigRequest
failed and messed up the RequestHandler.

Modifications
-------------
In the previous code, the message was failed, but instead the config
loading request should be given a chance to be retried (which the
upper logic does if an empty node array is returned) instead of
failing it prematurely.

Such a situation can occur when the node takes longer to go through
the full connect and auth cycle and may be available a few ms later.

Result
------
Graceful retry of configuration loading if the node is not (yet)
available.

Change-Id: Ia10b15ae515ac5f66e574d75d1234b7740a6255f
Reviewed-on: http://review.couchbase.org/42289
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>