JVMCBC-52: Upgrade RxJava to 1.0.0 RC8
Motivation
----------
In order to get RxJava closer to 1.0.0 GA, this upgrade makes sure
we are at the latest version available.

Modifications
-------------
The dependency was upgraded and both unit and integration tests
were run to make sure no existing functionality is broken.

Result
------
We are getting one of our core dependencies closer to 1.0.0 GA

Change-Id: Ic854cdd59ead47eaffa0e677a71310010e517a90
Reviewed-on: http://review.couchbase.org/42459
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>