JVMCBC-53: temporarly disable checkstyle in build
Motivation
----------
No Checkstyle configuration has been put in place yet (this is in the
works for 1.0.2), yet checkstyle is activated in the gradle build. This
causes a full build to fail (like in Jenkins).

Modifications
-------------
Disable checkstyle in gradle build until a configuration for it has
been made.

Result
------
Gradle doesn't fail because of checkstyle anymore.

Change-Id: I6666157632cf98d49546c3ba961d6a047c4fa622
Reviewed-on: http://review.couchbase.org/42652
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>