JVMCBC-55: Add reconnect delay ceiling for endpoints.
Motivation
----------
During failure tests it was discovered that the exponential backoff algorithm
used has no ceiling configured.

This is not a problem in lower bounds (millisecond range), but at one point
very quickly gets out of hand (4s, 8s, 16s,..) and potentially renders and
endpoint that is already up unusable for many seconds.

Modifications
-------------
A simple fixed ceiling at 4 seconds has been added so that a reconnect
attempt no longer waits than this time to reattempt the connect
process.

Result
------
A quicker and more reasonable approach to handle reconnect timings, even if it
takes a longer time to reconnect.

Change-Id: I2c836c7ebb7de4bad0572aa4e5ad08011bb198e4
Reviewed-on: http://review.couchbase.org/42702
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>