JVMCBC-62: Make Diagnostics resilient for not found classes
Motivation
----------
The code uses com/sun namespaced packages - which we shouldn't be using -
but are able to optionally utilize them to get more diagnostics information.

The code previously did check if the classes are proper instances, but it was
not resilient for environments where the class doesn't exist at all (that is
on IBM JVMs, as well as OSGi containers and app servers like Wildfly). It prevents
starting an application.

Modifications
-------------
The code now properly catches such an exception and logs a debug notice that extended
info is not available and only reduced output is printed.

Result
------
Properly degrading output and making sure app servers with reduced context are able
to start.

Change-Id: Ieb39fecdec77daf0f88ac8d757a9561c10da70a3
Reviewed-on: http://review.couchbase.org/42710
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>