JVMCBC-65: View & Query node locator only touches first node
Motivation
----------
Even if more nodes are available, because counter is never incremented, only
the first available node is tried to find a suitable node to write the request
to.

Modifications
-------------
The counter is now properly incremented, making sure all nodes are eventually
tried in a round-robin fashion as long as they are connected.

Result
------
Better distribution of the requests over the full cluster, leading to potentially
better performance and more equal resource utilization on the server side.

Change-Id: I85e5b0221e8b7c030ccdeb7b2d7c31aa1ad1d970
Reviewed-on: http://review.couchbase.org/43119
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>