JVMCBC-69: Do not let the RingBuffer handlers die under any circumstance.
Motivation
----------
The two core threads (the handlers for the Request and Response RingBuffers) are
very precious and should never die under any circumstance.

Modifications
-------------
Error handlers are attached to the Disruptor so that any exception that bubbles
up to this point (which is not intended, but can happen because of bugs), are
logged at the WARN level and need to be investigated, but the handlers itself
keep going and get notified of new events coming in.

Result
------
Better error debugging in case of unexpected exceptions, but also more resiliency
when they happen.

Change-Id: I9be70e4e51e31a550acc3e174d77d8224b914732
Reviewed-on: http://review.couchbase.org/43402
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>