JVMCBC-72: Use the right scheduler on bootstrap and increase logging.
Motivation
----------
It was discovered that some parts of the bootstrap process where using the
core and IO threads too much while they instead should be moved to the
computations scheduler earlier to free resources.

Increased logging has been added to the bootstrap process so that when
DEBUG is enabled, it is easer to find spots where threading is incorrect.

Modifications
-------------
Changed the schedulers early through observeOn, as well as increased
DEBUG logging verbosity throughout the bootstrap process.

Result
------
Quicker and less error prone bootstrapping, helping with debugging
of stuck bootstrap processes as well.

Change-Id: If167c0f9264cf18a1508a8c24514d917b76345e6
Reviewed-on: http://review.couchbase.org/43474
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>