JVMCBC-71: Move Handler Subjects onto different scheduler.
Motivation
----------
When Subjects are emitted from the IO threads, the code needs to make
sure that the following operators (from external code) are never executed
on the IO threads to avoid the risk of stalling.

Modifications
-------------
All exposed Subjects are now properly moved over to the computation scheduler
of the environment. As part of this effort, it was discovered through the test
cases that the subjects had problems with requested backpressure.

As suggested here (https://github.com/ReactiveX/RxJava/wiki/Backpressure), the code
also applies the onBackpressureBuffer operator.

Result
------
Consuming operators and subscribers are not able to block or in any other way mess
with the IO threads from where they are filled.

Change-Id: Ie51ba40550f6520674caad86019ca2ed24b55a34
Reviewed-on: http://review.couchbase.org/43477
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>