JVMCBC-84: Retain() instead of copy() in KeyValueHandler on content.
Motivation
----------
In the current codebase, the ByteBuf of the content body is copied before
passing to the higher levels. It turns out, this has been in place because
the netty decoder will free the original buffers.

This leads to higher TLAB allocation pressure than needed.

Modifications
-------------
By just retain()ining instead of copy()ing the content ByteBuf, we keep
control and avoid copying data around unnecessarily.

Result
------
Half the TLAB allocations, especially important on larger get requests
to reduce the work on the pool.

Change-Id: If4dbedb6a9022b5fd911feba1fbff4fc0da12b83
Reviewed-on: http://review.couchbase.org/44349
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>