JVMCBC-86: Do not manually cancel subscription on retry.
Motivation
----------
When an operation is scheduled for retry, the subscription should
not manually be removed. This can cause NPEs since its subject to
a race condition. In addition, this code is not known to bring
an actual benefit.

Modifications
-------------
Remove the manual subscription management and instead let the
scheduler handle it as expected.

Result
------
No potential NPE and more predictable behavior

Change-Id: I714e1701a8e34274460e02f323f559a6b7aa99f7
Reviewed-on: http://review.couchbase.org/44414
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>