JVMCBC-90: Avoid logging ClosedChannelExceptions as warnings
Motivation
----------
When a node is removed or failed over, the socket is closed and
the in-flight ops are cancelled. The upper retry logic can handle
retry, but the current endpoint code warns on the cancellation.

This is misleading, since the endpoint handles the scenario
gracefully.

Modifications
-------------
Explicitly exclude ClosedChannelException from WARN logging.

Result
------
Less verbose logging, not making users suspicious with "false
positive" warnings.

Change-Id: I1abab350d39f655353d7026a1ee2d6258f57934c
Reviewed-on: http://review.couchbase.org/44480
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>
Reviewed-on: http://review.couchbase.org/44690