JVMCBC-91: Add Delay classes to encapsulate behaviour.
Motivation
----------
In lots of places, just providing a delay and timeunit is not enough
and also limits flexibility. The Delay class, with its implementation
encapsulates delay behaviour in a reusable manner.

For example:

	Delay.linear(TimeUnit.SECONDS);

It should also be possible to define upper and lower bounds, as well
as multiplication factors to customize how the curve looks in its
various stages.

Modificiations
--------------
This changeset adds a Delay abstract class with factory methods and also
provides fixed, linear and exponential delays out of the box. It can
also be extended for custom delay implementations.

Note that the actual implementation constructors are package private, it
is required to use the Delay factory methods.

Result
------
Easy configurable and reusable delay components which expressive semantics.

Change-Id: I176b00ca88f6f09086656528b42cd8ab3a0c4bf6
Reviewed-on: http://review.couchbase.org/44510
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>