JVMCBC-92: Allow Observe interval to be configured.
Motivation
----------
As in the 1.* SDK, it should be possible to configure the interval
for the observe polling (ReplicateTo, PersistTo). Previously, it has
been fixed to 10ms, which can be quite high in different environments.

Modifications
-------------
This changeset changes the default Delay to exponential, starting at 10
microseconds and a maximum poll interval of 100 milliseconds. Making it
exponential rather than fixed allows for quick poll check on fast systems
and gracefully slow down on slower systems.

In addition, the delay is completely configurable through the environment.

Result
------
More flexible and configurable observe poll interval.

Change-Id: Icf9a0f6aeeda642097f173cbf9e0702eba390c14
Reviewed-on: http://review.couchbase.org/44521
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>