JVMCBC-57: DCP functionality
Motivation
----------
In order to be able to implement third-party application integrations,
such as message queues or external indexes, low-level API to DCP
protocol have to be implemented.

Modification
------------
Implemented DCPService, which is allows to listen to the data streams
from Couchbase.

Result
------
DCP implementation, which allows to experiment with DCP integrations.

Change-Id: I50bf350397f3bd6b4ce948212b7762ee0c243ce6
Reviewed-on: http://review.couchbase.org/43782
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Sergey Avseyev <sergey.avseyev@gmail.com>