JVMCBC-97: Do not try to parse View service when not available.
Motivation
----------
During a rebalance out scenario, when parsing the "old style config",
the server will remove the view service from the list when it is not
available anymore.

Not checking for null will result in a NPE.

Modifications
-------------
The code now checks if the view service is set in the config and only
exposes it if not null. This also automatically will trigger a view
service removal from the node down the reconfiguration paths.

Result
------
More robust behaviour during rebalance out.

Change-Id: I484cb39e99a48891de96531c48b20d61a6e9e9c8
Reviewed-on: http://review.couchbase.org/44719
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>