JVMCBC-98: Fall back to verbose config if terse returns 404
Motivation
----------
When the terse bucket config over HTTP is not accessible (which is the
case on older server versions), the code needs to fall back to the
verbose config when it both fails or just returns a 404 not found.

Modifications
-------------
The code now properly checks if the terse config is a success and if not,
converts it into an exception that will be caught by the next handler.
Note that while this may not be the most performant way to handle this,
the code path is not travelled much and makes error handling more uniform.

Also, a test case has been added and another one corrected (the wrong call
was used and overriden by the success message, leading to an always
succeeding test).

Result
------
Correct fallback handling on older server versions.

Change-Id: I1e61ff30599dbac153da4a65d15334683918436b
Reviewed-on: http://review.couchbase.org/44748
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>
Reviewed-on: http://review.couchbase.org/44768