JVMCBC-99: Do not propose invalid config in Carrier Refresher.
Motivation
----------
The refresher needs to make sure that only valid configs are proposed to the the
configuration provider. Especially when the polling happens against an older
cluster, the command might not return a successful JSON configuration.

Modifications
-------------
The code now properly adapts for non-success responses and also checks if the
content is - at least simply - JSON.

Result
------
More robust handling against older clusters.

Change-Id: I0811019a1291e7a703864918b7a63d41beac2f52
Reviewed-on: http://review.couchbase.org/44751
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>
Reviewed-on: http://review.couchbase.org/44770