JVMCBC-100: Take the first good config on loading, skip the others.
Motivation
----------
When a list of good bootstrap nodes is passed in, the code fans out
to grab a config from all of them. While this itself is valid, all
of those configs are passed up to the configuration provider and
are applied.

This can lead to double "bucket opened" messages and also duplicate
work to be performed.

Modifications
-------------
Only take the first successful configuration and then unsubscribe.

Result
------
Less work to be done in the configuration provider and no misleading
info messages for the user.

Change-Id: I66440db95843aac75d412e46ccc06c447bd376bf
Reviewed-on: http://review.couchbase.org/44754
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>