Set ByteBuf leak detector to paranoid in integration tests.
Motivation
----------
In order to more quickly detect ByteBuf leaks, all integration tests
should be run with the leak detector switched on and in the "paranoid"
setting to detect as many leaks as possible.

Modifications
-------------
Since the ClusterDependentTest is the parent class for all integration
tests, the static property is set there.

Result
------
Potentially better results in early identifying ByteBuf pool leaks.

Change-Id: I06754196171352b391e8bea67e00cb3e1d3326e1
Reviewed-on: http://review.couchbase.org/44761
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>