JVMCBC-104: CouchbaseNodes need to be equals on same InetAddr
Motivation
----------
The RequestHandler stores all the CouchbaseNodes in a Set. Even
when race conditions happen, we need to make sure that the list
of nodes never contains duplicates.

Modifications
-------------
The CouchbaseNode now has custom hashCode and equals methods which
align on the stored InetAddress. That way, nodes pointing to the
same address are treated the same and never stored together in the
set.

Result
------
More consistent handling of nodes, and never duplicates in the
node list set.

Change-Id: I23cbdd72f655870777bc3476e779da8c0822f40d
Reviewed-on: http://review.couchbase.org/44763
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>