Narrowing misleading DEBUG message for node removal.
Motivation
----------
The DEBUG message showed a list of node removal _candidates_, but
logged them as they would be removed.

Modifications
-------------
Only log a DEBUG message if the node is actually going to be removed.

Result
------
No misleading DEBUG log message on node removal if its not actually
going to happen.

Change-Id: I60e6e7d0e2b2820f1a98f62b0ca66b326751cc4b
Reviewed-on: http://review.couchbase.org/44764
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>
Reviewed-on: http://review.couchbase.org/44775